requirements/core.txt
# Core Framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8
PyJWT==2.10.1

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.19.1
cerberus==1.3.5

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Web and Communication
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
websockets==12.0 